move home feed components into separate files
