+++ apple/OmnivoreKit/Sources/App/Views/Home/Components/FeedItemContextMenuView.swift
@@ -42,42 +42,3 @@ struct FeedItemContextMenuView: View {
     }
   }
 }
-
-struct FeedCardNavigationLink: View {
-  @EnvironmentObject var dataService: DataService
-
-  let item: FeedItem
-  let searchQuery: String
-
-  @Binding var selectedLinkItem: FeedItem?
-
-  @ObservedObject var viewModel: HomeFeedViewModel
-  var body: some View {
-    NavigationLink(
-      destination: LinkItemDetailView(viewModel: LinkItemDetailViewModel(item: item)),
-      tag: item,
-      selection: $selectedLinkItem
-    ) {
-      EmptyView()
-    }
-    .opacity(0)
-    .buttonStyle(PlainButtonStyle())
-    .onAppear {
-      viewModel.itemAppeared(item: item, searchQuery: searchQuery, dataService: dataService)
-    }
-    FeedCard(item: item)
-  }
-}
-
-struct LoadingSection: View {
-  var body: some View {
-    Section {
-      HStack(alignment: .center) {
-        Spacer()
-        Text("Loading...")
-        Spacer()
-      }
-      .frame(maxWidth: .infinity)
-    }
-  }
-}
